server/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
psycopg2-binary
bcrypt
python-dotenv
flask-migrate
flask-login
marshmallow
pandas